+++ lib/src/helpers/helpers.dart
@@ -1,2 +1,1 @@
-export 'console.dart';
 export 'localization_helpers.dart';